+++ frontend/src/components/layout/TabBar/TabBar.tsx
@@ -1,59 +1,11 @@
-import styled from '@emotion/styled';
-import { FaPencilAlt, FaUser } from 'react-icons/fa';
-import { HiMenu } from 'react-icons/hi';
-
-export const Section = styled.section`
-	position: sticky;
-	bottom: 0;
-	width: 100%;
-	height: 3.5rem;
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	box-shadow: 0px -4px 15px ${({ theme }) => theme.boxShadows.secondary};
-	border-radius: ${({ theme }) => theme.fonts.SIZE_010} ${({ theme }) => theme.fonts.SIZE_010} 0 0;
-	z-index: ${({ theme }) => theme.zIndex.HEADER};
-	background-color: ${({ theme }) => theme.colors.WHITE};
-`;
-
-export const PostingLink = styled(FaPencilAlt)`
-	font-size: ${({ theme }) => theme.fonts.SIZE_022};
-	color: ${({ theme }) => theme.colors.BLACK_500};
-	cursor: pointer;
-	margin: 16px 0 16px 16px;
-	&:hover,
-	&:active {
-		color: ${({ theme }) => theme.colors.PURPLE_500};
-	}
-`;
-
-export const UserCircleLink = styled(FaUser)`
-	font-size: ${({ theme }) => theme.fonts.SIZE_022};
-	color: ${({ theme }) => theme.colors.BLACK_500};
-	cursor: pointer;
-	&:hover,
-	&:active {
-		color: ${({ theme }) => theme.colors.PURPLE_500};
-	}
-`;
-
-export const MenuLink = styled(HiMenu)`
-	font-size: ${({ theme }) => theme.fonts.SIZE_024};
-	color: ${({ theme }) => theme.colors.BLACK_500};
-	cursor: pointer;
-	margin: 16px 16px 16px 0;
-	&:hover,
-	&:active {
-		color: ${({ theme }) => theme.colors.PURPLE_500};
-	}
-`;
+import * as S from '@/components/layout/TabBar/TabBat.style';
 
 const TabBar = () => (
-	<Section>
-		<PostingLink />
-		<UserCircleLink />
-		<MenuLink />
-	</Section>
+	<S.Section>
+		<S.PostingLink />
+		<S.UserCircleLink />
+		<S.MenuLink />
+	</S.Section>
 );
 
 export default TabBar;
