+++ frontend/src/pages/Login/index.style.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+	display: flex;
+	width: 100%;
+	height: calc(100vh - 15rem);
+	justify-content: center;
+	align-items: center;
+`;
